requirements.txt
streamlit
pdfplumber
plotly
pandas
fpdf2
requests